requirements.txt
PyMuPDF
Pillow
fastapi>=0.110.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
aiofiles>=24.1.0
requests>=2.32.3
python-multipart>=0.0.9
